pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "opencv-python>=4.11.0.86",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.37",
    "werkzeug>=3.1.3",
    "numpy>=2.2.2",
    "opencv-python-headless>=4.11.0.86",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
]
